pkg/api/core: move pkg/api/service
